LB-2298: Apply https://github.com/aws/aws-sdk-java/issues/427 recommendation.
According to
https://github.com/aws/aws-sdk-java/issues/427#issuecomment-100518891 and
https://github.com/aws/aws-sdk-java/issues/427#issuecomment-100583279, since
we don't use a File or FileInputStream directly (stream here is either
CipherWithInlineIVInputStream -if we encrypt- or BufferedInputStream) we
should set the "read limit" (which is the maximum buffer size that could be
consumed) as suggested in the error message mentioned in LB-2298. That limit
should be the expected max size of our input stream in bytes, plus 1, hence
partSize+1.