Made upload command/methods path/URI handling more consistent
--recursive flag was introduced to make it more explicit when a directory is
going to be uploaded and, also, make the command more similar to the rest of
the CRUD commands (e.g. download, copy). --recursive requires a directory-like
destination URI and, currently, it cannot be combined with non-directory
inputs.

This change is reflected to the API as well. The `uploadDirectory` method was
renamed to `uploadRecursively` but the semantics remained the same.