Made delete command/methods URI handling more consistent
At the CLI level, the delete command now expects either a fully qualified
file URI or a prefix URI + --recursive. A directory-like URI (i.e. ending with
‘/’) is considered a prefix as well and, hence, the “--recursive” option will
be required.

This change is reflected to the API as well. The `delete` method expects a
fully qualified file URI, while the `deleteRecursively` expects a prefix and
it's going to delete all files staring with that prefix.

The `deleteDirectory` method has been renamed to `deleteRecursively` to match
its purpose more accurately.